# Conflicts:
#	SMPTrustedServiceManager/pom.xml
#	pom.xml

+++ src/main/java/com/apple/iossystems/smp/reporting/core/timer/Timer.java
@@ -9,7 +9,7 @@ public class Timer
     {
     }
 
-    public static boolean delayExpired(long timeInMilliseconds, int delay)
+    public static boolean delayExpired(long timeInMilliseconds, long delay)
     {
         return ((System.currentTimeMillis() - timeInMilliseconds) >= delay);
     }
